AsteroidBug temp fix

+++ Assets/Scripts/ScriptableObjects/SpaceUnitConfig.cs
@@ -7,8 +7,8 @@ namespace Config
     public class SpaceUnitConfig : ScriptableObject {
         public float maxAngleSpeed;
         public float maxSpeed;
-        public int maxHp;
-        public int currentHp;
+        public float maxHp;
+        public float currentHp;
         public bool isDockable;
         public Vector3 position = Vector3.one;
         public Quaternion rotation = Quaternion.identity;
